+++ core/container.go
@@ -8,7 +8,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/docker/distribution/reference"
 	bkclient "github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/moby/buildkit/exporter/containerimage/exptypes"
@@ -16,9 +15,7 @@ import (
 	"github.com/moby/buildkit/solver/pb"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
-	"go.dagger.io/dagger/core/schema"
 	"go.dagger.io/dagger/core/shim"
-	"go.dagger.io/dagger/router"
 )
 
 // Container is a content-addressed container.
@@ -659,367 +656,8 @@ func (container *Container) Publish(
 	return ref, nil
 }
 
-type containerSchema struct {
-	*baseSchema
-}
-
-var _ router.ExecutableSchema = &containerSchema{}
-
-func (s *containerSchema) Name() string {
-	return "container"
-}
-
-func (s *containerSchema) Schema() string {
-	return schema.Container
-}
-
-func (s *containerSchema) Resolvers() router.Resolvers {
-	return router.Resolvers{
-		"ContainerID":      stringResolver(ContainerID("")),
-		"ContainerAddress": stringResolver(ContainerAddress("")),
-		"Query": router.ObjectResolver{
-			"container": router.ToResolver(s.container),
-		},
-		"Container": router.ObjectResolver{
-			"from":                 router.ToResolver(s.from),
-			"rootfs":               router.ToResolver(s.rootfs),
-			"directory":            router.ToResolver(s.directory),
-			"user":                 router.ToResolver(s.user),
-			"withUser":             router.ToResolver(s.withUser),
-			"workdir":              router.ToResolver(s.workdir),
-			"withWorkdir":          router.ToResolver(s.withWorkdir),
-			"variables":            router.ToResolver(s.variables),
-			"variable":             router.ToResolver(s.variable),
-			"withVariable":         router.ToResolver(s.withVariable),
-			"withSecretVariable":   router.ErrResolver(ErrNotImplementedYet),
-			"withoutVariable":      router.ToResolver(s.withoutVariable),
-			"entrypoint":           router.ToResolver(s.entrypoint),
-			"withEntrypoint":       router.ToResolver(s.withEntrypoint),
-			"defaultArgs":          router.ToResolver(s.defaultArgs),
-			"withDefaultArgs":      router.ToResolver(s.withDefaultArgs),
-			"mounts":               router.ToResolver(s.mounts),
-			"withMountedDirectory": router.ToResolver(s.withMountedDirectory),
-			"withMountedFile":      router.ToResolver(s.withMountedFile),
-			"withMountedTemp":      router.ToResolver(s.withMountedTemp),
-			"withMountedCache":     router.ToResolver(s.withMountedCache),
-			"withMountedSecret":    router.ErrResolver(ErrNotImplementedYet),
-			"withoutMount":         router.ToResolver(s.withoutMount),
-			"exec":                 router.ToResolver(s.exec),
-			"exitCode":             router.ToResolver(s.exitCode),
-			"stdout":               router.ToResolver(s.stdout),
-			"stderr":               router.ToResolver(s.stderr),
-			"publish":              router.ToResolver(s.publish),
-		},
-	}
-}
-
-func (s *containerSchema) Dependencies() []router.ExecutableSchema {
-	return nil
-}
-
-type containerArgs struct {
-	ID ContainerID
-}
-
-func (s *containerSchema) container(ctx *router.Context, parent any, args containerArgs) (*Container, error) {
-	return &Container{
-		ID: args.ID,
-	}, nil
-}
-
-type containerFromArgs struct {
-	Address ContainerAddress
-}
-
-func (s *containerSchema) from(ctx *router.Context, parent *Container, args containerFromArgs) (*Container, error) {
-	addr := string(args.Address)
-
-	refName, err := reference.ParseNormalizedNamed(addr)
-	if err != nil {
-		return nil, err
-	}
-
-	ref := reference.TagNameOnly(refName).String()
-
-	_, cfgBytes, err := s.gw.ResolveImageConfig(ctx, ref, llb.ResolveImageConfigOpt{
-		Platform:    &s.platform,
-		ResolveMode: llb.ResolveModeDefault.String(),
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	var imgSpec specs.Image
-	if err := json.Unmarshal(cfgBytes, &imgSpec); err != nil {
-		return nil, err
-	}
-
-	ctr, err := parent.WithFS(ctx, llb.Image(addr), s.platform)
-	if err != nil {
-		return nil, err
-	}
-
-	return ctr.UpdateImageConfig(ctx, func(specs.ImageConfig) specs.ImageConfig {
-		return imgSpec.Config
-	})
-}
-
-func (s *containerSchema) rootfs(ctx *router.Context, parent *Container, args any) (*Directory, error) {
-	return parent.FS(ctx)
-}
-
-type containerExecArgs struct {
-	// Args is optional. If it is nil, we the default args for the image.
-	Args *[]string
-	Opts ContainerExecOpts
-}
-
 type ContainerExecOpts struct {
 	Stdin          *string
 	RedirectStdout *string
 	RedirectStderr *string
-}
-
-func (s *containerSchema) exec(ctx *router.Context, parent *Container, args containerExecArgs) (*Container, error) {
-	return parent.Exec(ctx, s.gw, args.Args, args.Opts)
-}
-
-func (s *containerSchema) exitCode(ctx *router.Context, parent *Container, args any) (*int, error) {
-	return parent.ExitCode(ctx, s.gw)
-}
-
-func (s *containerSchema) stdout(ctx *router.Context, parent *Container, args any) (*File, error) {
-	return parent.MetaFile(ctx, s.gw, "stdout")
-}
-
-func (s *containerSchema) stderr(ctx *router.Context, parent *Container, args any) (*File, error) {
-	return parent.MetaFile(ctx, s.gw, "stderr")
-}
-
-type containerWithEntrypointArgs struct {
-	Args []string
-}
-
-func (s *containerSchema) withEntrypoint(ctx *router.Context, parent *Container, args containerWithEntrypointArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		cfg.Entrypoint = args.Args
-		return cfg
-	})
-}
-
-func (s *containerSchema) entrypoint(ctx *router.Context, parent *Container, args containerWithVariableArgs) ([]string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return cfg.Entrypoint, nil
-}
-
-type containerWithDefaultArgs struct {
-	Args *[]string
-}
-
-func (s *containerSchema) withDefaultArgs(ctx *router.Context, parent *Container, args containerWithDefaultArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		if args.Args == nil {
-			cfg.Cmd = []string{}
-			return cfg
-		}
-
-		cfg.Cmd = *args.Args
-		return cfg
-	})
-}
-
-func (s *containerSchema) defaultArgs(ctx *router.Context, parent *Container, args any) ([]string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return cfg.Cmd, nil
-}
-
-type containerWithUserArgs struct {
-	Name string
-}
-
-func (s *containerSchema) withUser(ctx *router.Context, parent *Container, args containerWithUserArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		cfg.User = args.Name
-		return cfg
-	})
-}
-
-func (s *containerSchema) user(ctx *router.Context, parent *Container, args containerWithVariableArgs) (string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return "", err
-	}
-
-	return cfg.User, nil
-}
-
-type containerWithWorkdirArgs struct {
-	Path string
-}
-
-func (s *containerSchema) withWorkdir(ctx *router.Context, parent *Container, args containerWithWorkdirArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		cfg.WorkingDir = absPath(cfg.WorkingDir, args.Path)
-		return cfg
-	})
-}
-
-func (s *containerSchema) workdir(ctx *router.Context, parent *Container, args containerWithVariableArgs) (string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return "", err
-	}
-
-	return cfg.WorkingDir, nil
-}
-
-type containerWithVariableArgs struct {
-	Name  string
-	Value string
-}
-
-func (s *containerSchema) withVariable(ctx *router.Context, parent *Container, args containerWithVariableArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		// NB(vito): buildkit handles replacing properly when we do llb.AddEnv, but
-		// we want to replace it here anyway because someone might publish the image
-		// instead of running it. (there's a test covering this!)
-		newEnv := []string{}
-		prefix := args.Name + "="
-		for _, env := range cfg.Env {
-			if !strings.HasPrefix(env, prefix) {
-				newEnv = append(newEnv, env)
-			}
-		}
-
-		newEnv = append(newEnv, fmt.Sprintf("%s=%s", args.Name, args.Value))
-
-		cfg.Env = newEnv
-
-		return cfg
-	})
-}
-
-type containerWithoutVariableArgs struct {
-	Name string
-}
-
-func (s *containerSchema) withoutVariable(ctx *router.Context, parent *Container, args containerWithoutVariableArgs) (*Container, error) {
-	return parent.UpdateImageConfig(ctx, func(cfg specs.ImageConfig) specs.ImageConfig {
-		removedEnv := []string{}
-		prefix := args.Name + "="
-		for _, env := range cfg.Env {
-			if !strings.HasPrefix(env, prefix) {
-				removedEnv = append(removedEnv, env)
-			}
-		}
-
-		cfg.Env = removedEnv
-
-		return cfg
-	})
-}
-
-func (s *containerSchema) variables(ctx *router.Context, parent *Container, args any) ([]string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	return cfg.Env, nil
-}
-
-type containerVariableArgs struct {
-	Name string
-}
-
-func (s *containerSchema) variable(ctx *router.Context, parent *Container, args containerVariableArgs) (*string, error) {
-	cfg, err := parent.ImageConfig(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, env := range cfg.Env {
-		name, val, ok := strings.Cut(env, "=")
-		if ok && name == args.Name {
-			return &val, nil
-		}
-	}
-
-	return nil, nil
-}
-
-type containerWithMountedDirectoryArgs struct {
-	Path   string
-	Source DirectoryID
-}
-
-func (s *containerSchema) withMountedDirectory(ctx *router.Context, parent *Container, args containerWithMountedDirectoryArgs) (*Container, error) {
-	return parent.WithMountedDirectory(ctx, args.Path, &Directory{ID: args.Source})
-}
-
-type containerPublishArgs struct {
-	Address ContainerAddress
-}
-
-func (s *containerSchema) publish(ctx *router.Context, parent *Container, args containerPublishArgs) (ContainerAddress, error) {
-	return parent.Publish(ctx, args.Address, s.bkClient, s.solveOpts, s.solveCh)
-}
-
-type containerWithMountedFileArgs struct {
-	Path   string
-	Source FileID
-}
-
-func (s *containerSchema) withMountedFile(ctx *router.Context, parent *Container, args containerWithMountedFileArgs) (*Container, error) {
-	return parent.WithMountedFile(ctx, args.Path, &File{ID: args.Source})
-}
-
-type containerWithMountedCacheArgs struct {
-	Path   string
-	Source DirectoryID
-}
-
-func (s *containerSchema) withMountedCache(ctx *router.Context, parent *Container, args containerWithMountedCacheArgs) (*Container, error) {
-	var dir *Directory
-	if args.Source != "" {
-		dir = &Directory{ID: args.Source}
-	}
-
-	return parent.WithMountedCache(ctx, args.Path, dir)
-}
-
-type containerWithMountedTempArgs struct {
-	Path string
-}
-
-func (s *containerSchema) withMountedTemp(ctx *router.Context, parent *Container, args containerWithMountedTempArgs) (*Container, error) {
-	return parent.WithMountedTemp(ctx, args.Path)
-}
-
-type containerWithoutMountArgs struct {
-	Path string
-}
-
-func (s *containerSchema) withoutMount(ctx *router.Context, parent *Container, args containerWithoutMountArgs) (*Container, error) {
-	return parent.WithoutMount(ctx, args.Path)
-}
-
-func (s *containerSchema) mounts(ctx *router.Context, parent *Container, _ any) ([]string, error) {
-	return parent.Mounts(ctx)
-}
-
-type containerDirectoryArgs struct {
-	Path string
-}
-
-func (s *containerSchema) directory(ctx *router.Context, parent *Container, args containerDirectoryArgs) (*Directory, error) {
-	return parent.Directory(ctx, s.gw, args.Path)
 }
