+++ core/util.go
@@ -5,29 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"path"
-	"strings"
 
-	"github.com/graphql-go/graphql/language/ast"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/moby/buildkit/solver/pb"
-	"github.com/pkg/errors"
-	"go.dagger.io/dagger/router"
 )
-
-// ErrNotImplementedYet is used to stub out API fields that aren't implemented
-// yet.
-var ErrNotImplementedYet = errors.New("not implemented yet")
-
-func truncate(s string, lines *int) string {
-	if lines == nil {
-		return s
-	}
-	l := strings.SplitN(s, "\n", *lines+1)
-	if *lines > len(l) {
-		*lines = len(l)
-	}
-	return strings.Join(l[0:*lines], "\n")
-}
 
 // encodeID JSON marshals and base64-encodes an arbitrary payload.
 func encodeID(payload any) (string, error) {
@@ -55,34 +36,16 @@ func decodeID[T ~string](payload any, id T) error {
 	return json.Unmarshal(jsonBytes, payload)
 }
 
-// stringResolver is used to generate a scalar resolver for a stringable type.
-func stringResolver[T ~string](sample T) router.ScalarResolver {
-	return router.ScalarResolver{
-		Serialize: func(value any) any {
-			switch v := value.(type) {
-			case string, T:
-				return v
-			default:
-				panic(fmt.Sprintf("unexpected %T type %T", sample, v))
-			}
-		},
-		ParseValue: func(value any) any {
-			switch v := value.(type) {
-			case string:
-				return T(v)
-			default:
-				panic(fmt.Sprintf("unexpected %T value type %T: %+v", sample, v, v))
-			}
-		},
-		ParseLiteral: func(valueAST ast.Value) any {
-			switch valueAST := valueAST.(type) {
-			case *ast.StringValue:
-				return T(valueAST.Value)
-			default:
-				panic(fmt.Sprintf("unexpected %T literal type: %T", sample, valueAST))
-			}
-		},
+func absPath(workDir string, containerPath string) string {
+	if path.IsAbs(containerPath) {
+		return containerPath
 	}
+
+	if workDir == "" {
+		workDir = "/"
+	}
+
+	return path.Join(workDir, containerPath)
 }
 
 func defToState(def *pb.Definition) (llb.State, error) {
@@ -100,16 +63,4 @@ func defToState(def *pb.Definition) (llb.State, error) {
 	}
 
 	return llb.NewState(defop), nil
-}
-
-func absPath(workDir string, containerPath string) string {
-	if path.IsAbs(containerPath) {
-		return containerPath
-	}
-
-	if workDir == "" {
-		workDir = "/"
-	}
-
-	return path.Join(workDir, containerPath)
 }
